requirements.txt
django
Pillow
django_cleanup
django-ckeditor